+++ ide/MoonshineSharedCore/src/elements/chrome-debug-adapter/out/src/chromeProvidedPortConnection.js
@@ -4,14 +4,16 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
         function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.ChromeProvidedPortConnection = void 0;
 const vscode_chrome_debug_core_1 = require("vscode-chrome-debug-core");
 const vscode_chrome_debug_core_2 = require("vscode-chrome-debug-core");
 const vscode_chrome_debug_core_3 = require("vscode-chrome-debug-core");